remove: useChangePrivate 삭제 후 폴더에 내장
