+++ front/src/hooks/index.tsx
@@ -1,2 +1,1 @@
 export { default as useLogout } from './useLogout';
-export { default as useChangePrivate } from './useChangePrivate';
